+++ src/parser.rs
@@ -1,23 +1,92 @@
 use crate::tag::Tag;
 
 pub const RESET_FOREGROUND_CLASS: &str = "icolor";
-const SPOILER_START: &str = "[spoiler]";
-const SPOILER_END: &str = "[/spoiler]";
+const SPOILER_OPEN_TAG: &str = "[spoiler]";
+const SPOILER_CLOSE_TAG: &str = "[/spoiler]";
 
-macro_rules! close_spoiler {
-    ($idx:expr, $spoiler_tag:ident, $paragraph_tag:ident, $line:ident) => {
-        // Avoid appending empty spoilers
-        if $idx != 0 {
-            $spoiler_tag.child(&$line[..$idx]);
-            $paragraph_tag.child($spoiler_tag);
-        }
+#[derive(Debug)]
+enum Token {
+    SpoilerOpen,
+    SpoilerClose,
+    Text(String),
+}
 
-        // The spoiler could be closing at the end of the line, avoid appending it to the paragraph
-        // on that case
-        $line = &$line[$idx + SPOILER_END.len()..];
-        if !$line.is_empty() {
-            $paragraph_tag.child($line);
+macro_rules! tokenize_spoiler {
+    ($ref:expr, $idx:expr, $tokens:expr, $line:ident, $token_kind:ident) => {
+        // Push the text before the tag if any.
+        if $idx != $ref {
+            $tokens.push(Token::Text($line[$ref..$idx].into()));
         }
+        $tokens.push(Token::$token_kind);
+    };
+    ($idx:expr, $tokens:expr, $line:ident, $token_kind:ident) => {
+        tokenize_spoiler!(0, $idx, $tokens, $line, $token_kind);
+    };
+}
+
+macro_rules! push_line_buffer {
+    ($line:ident, $idx:expr, $tag:expr) => {
+        $line = &$line[$idx + $tag.len()..];
+    };
+}
+
+fn tokenize<S: ?Sized>(line: &S) -> Vec<Token>
+where
+    S: AsRef<str>,
+{
+    let mut line = line.as_ref();
+    let mut tokens: Vec<Token> = Vec::new();
+
+    loop {
+        match (line.find(SPOILER_OPEN_TAG), line.find(SPOILER_CLOSE_TAG)) {
+            (Some(start_idx), Some(end_idx)) => {
+                // While both tags are in the same line, they could be in any order.
+                if start_idx < end_idx {
+                    tokenize_spoiler!(start_idx, tokens, line, SpoilerOpen);
+                    tokenize_spoiler!(
+                        start_idx + SPOILER_OPEN_TAG.len(),
+                        end_idx,
+                        tokens,
+                        line,
+                        SpoilerClose
+                    );
+                    push_line_buffer!(line, end_idx, SPOILER_CLOSE_TAG);
+                } else {
+                    tokenize_spoiler!(end_idx, tokens, line, SpoilerClose);
+                    tokenize_spoiler!(
+                        end_idx + SPOILER_CLOSE_TAG.len(),
+                        start_idx,
+                        tokens,
+                        line,
+                        SpoilerOpen
+                    );
+                    push_line_buffer!(line, start_idx, SPOILER_OPEN_TAG);
+                }
+            }
+            (Some(idx), None) => {
+                tokenize_spoiler!(idx, tokens, line, SpoilerOpen);
+                push_line_buffer!(line, idx, SPOILER_OPEN_TAG);
+            }
+            (None, Some(idx)) => {
+                tokenize_spoiler!(idx, tokens, line, SpoilerClose);
+                push_line_buffer!(line, idx, SPOILER_CLOSE_TAG);
+            }
+            (None, None) => break,
+        };
+    }
+
+    if !line.is_empty() {
+        tokens.push(Token::Text(line.into()))
+    }
+
+    tokens
+}
+
+macro_rules! push_span_tag {
+    ($paragraph:ident, $text:expr) => {
+        let mut spoiler = Tag::new("span");
+        spoiler.child($text);
+        $paragraph.child(spoiler);
     };
 }
 
@@ -31,63 +100,45 @@ impl LineParser {
         self.open_spoiler
     }
 
+    // FIX: Merge sequential span tags together.
     pub fn parse<S: ?Sized>(&mut self, line: &S) -> Tag
     where
         S: AsRef<str>,
     {
-        let line = line.as_ref();
+        let mut paragraph = Tag::new("p");
+        let mut is_first_text = true;
 
-        let mut paragraph_tag = Tag::new("p");
-        let mut spoiler_tag = Tag::new("span");
-
-        // Remove highlight if it doesn't apply to the given line
-        let trimmed_line = line.trim_start();
-        if !trimmed_line.starts_with('>')
-            && !trimmed_line.starts_with(&format!("{}>", SPOILER_START))
-            && !trimmed_line.starts_with(&format!("{}>", SPOILER_END))
-        {
-            paragraph_tag.attribute("class", RESET_FOREGROUND_CLASS);
-        }
-
-        // Parse spoilers
-        if self.open_spoiler {
-            let mut line = line;
-
-            if let Some(idx) = line.find(SPOILER_END) {
-                close_spoiler!(idx, spoiler_tag, paragraph_tag, line);
-                self.open_spoiler = false;
-            } else {
-                // Here we are in a line that doesn't close the spoiler coming from lines back
-                spoiler_tag.child(line);
-                paragraph_tag.child(spoiler_tag);
-            }
-        } else if let Some(idx) = line.find(SPOILER_START) {
-            let mut line = line;
-
-            // Avoiding appending an empty text child at the start of the paragraph for spoilers at
-            // the start of the line
-            if idx != 0 {
-                paragraph_tag.child(&line[..idx]);
-            }
-
-            line = &line[idx + SPOILER_START.len()..];
-            if let Some(idx) = line.find(SPOILER_END) {
-                close_spoiler!(idx, spoiler_tag, paragraph_tag, line);
-            } else {
-                // Here we are in a line with an unclosed spoiler, which could be starting at the
-                // end of the line. Avoid appending on that case for this line.
-                if !line.is_empty() {
-                    spoiler_tag.child(line);
-                    paragraph_tag.child(spoiler_tag);
+        for token in tokenize(line) {
+            // Remove highlight if it doesn't apply to the given line
+            if is_first_text {
+                if let Token::Text(text) = &token {
+                    if !text.starts_with('>') {
+                        paragraph.attribute("class", RESET_FOREGROUND_CLASS);
+                    }
+                    is_first_text = false;
                 }
-
-                self.open_spoiler = true;
             }
-        } else {
-            paragraph_tag.child(line);
+
+            match token {
+                Token::Text(text) if self.open_spoiler => {
+                    push_span_tag!(paragraph, text);
+                }
+                Token::Text(text) => {
+                    paragraph.child(text);
+                }
+                // Place the tag as is if `open_spoiler` would stay the same.
+                Token::SpoilerOpen if self.open_spoiler => {
+                    push_span_tag!(paragraph, SPOILER_OPEN_TAG);
+                }
+                Token::SpoilerClose if !self.open_spoiler => {
+                    push_span_tag!(paragraph, SPOILER_CLOSE_TAG);
+                }
+                Token::SpoilerOpen => self.open_spoiler = true,
+                Token::SpoilerClose => self.open_spoiler = false,
+            };
         }
 
-        paragraph_tag
+        paragraph
     }
 }
 
@@ -290,6 +341,59 @@ mod test {
             parser,
             tag!(hi, ">Nothing here either"),
             "[/spoiler]>Nothing here either"
+        );
+    }
+
+    #[test]
+    fn several_spoilers_in_a_single_line() {
+        let mut parser = LineParser::default();
+
+        assert_parse!(
+            parser,
+            tag!(
+                spoiler!("This is a paragraph"),
+                " with ",
+                spoiler!("two spoilers")
+            ),
+            "[spoiler]This is a paragraph[/spoiler] with [spoiler]two spoilers[/spoiler]"
+        );
+        assert_parse!(
+            parser,
+            tag!(spoiler!("This is"), " a ", spoiler!("paragraph"), " with ", spoiler!("three spoilers")),
+            "[spoiler]This is[/spoiler] a [spoiler]paragraph[/spoiler] with [spoiler]three spoilers[/spoiler]"
+        );
+        assert_parse!(
+            parser,
+            tag!(spoiler!("This is"), " a ", spoiler!("paragraph"), " with ", spoiler!("four"), " ", spoiler!("spoilers")),
+            "[spoiler]This is[/spoiler] a [spoiler]paragraph[/spoiler] with [spoiler]four[/spoiler] [spoiler]spoilers[/spoiler]"
+        );
+    }
+
+    #[test]
+    fn several_spoilers_multiple_lines() {
+        let mut parser = LineParser::default();
+
+        assert_parse!(
+            spoiler,
+            parser,
+            tag!(
+                "The ",
+                spoiler!("spoiler"),
+                " starts on ",
+                spoiler!("this line")
+            ),
+            "The [spoiler]spoiler[/spoiler] starts on [spoiler]this line"
+        );
+        assert_parse!(
+            spoiler,
+            parser,
+            tag!(spoiler!("then ends on the following one"), ", next is ", spoiler!("a middle one"), " and then ", spoiler!("it starts again")),
+            "then ends on the following one[/spoiler], next is [spoiler]a middle one[/spoiler] and then [spoiler]it starts again"
+        );
+        assert_parse!(
+            parser,
+            tag!(spoiler!("to end"), " and never come back again"),
+            "to end[/spoiler] and never come back again"
         );
     }