+++ CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for multiple spoilers in the same line.
+
 ### Changed
 - Avoid creating an EPUB without any text files to convert.
 - Check if `COLOR` argument on `--green-color` and `--spoiler-color`